Post-Its Are Working!
Some more environment work, post-its work now, etc. Things I still need
to do include (but are not limited to): more lighting, more environment
work, put in the door for the end of the game, get record player and
(hopefully TV) working, get a post-it model to use, sounds, and get all
the post its in game and triggering to move the story along. Woo...
GAHH.