Dedicated Version
I added a fading bit of text when the house scene loads that contains a
dedication to Jasen and Gus. This was done upon Jeremy's request because
he simply wanted to fuck with Jasen. I kind of had to. The text is
turned off in the game, but I built it with it enabled once for Jeremy.